Split sky_params data by consumer; move ramp types to util
The single common::sky_params module mixed the reusable ramp types with
data owned by two different consumers. Split it:

- common::util gains the generic Gradient/Curve ramp types (not sky-specific)
- scene_runner owns the light/ambient/fog gradients (light_gradients)
- visuals owns the atmosphere rayleigh/mie/night coefficients (atmosphere_params)

No behaviour change; common no longer carries the sky data, only the shared type.
